correccion para pruebas

+++ src/pages/Profesores.jsx
@@ -6,25 +6,48 @@ import useLogin from "../hooks/useLogin";
 import { Button } from "react-bootstrap";
 import moment from "moment";
 import { useNavigate } from "react-router-dom";
+import ToastBootstrap from "../components/Toasts";
 
 const Profesores = () => {
     const [profesores, setProfesores] = useState()
+    const [showToast, setShowToast] = useState(false)
+    const [toastMessage, setToastMessage] = useState({
+      title: "",
+      message: "",
+      color:""
+    });
     const { auth } = useLogin()
     const navigate = useNavigate();
-
-    const eliminarUsuario = async (id) => {
+    const obtenerUsuarios = async () => {
       try {
-          const response = await axios.delete(`http://localhost:3000/usuarios/${id}`, {
+        const response = await axios.get('http://localhost:3000/usuarios?rol=PROFESOR', {
+          headers: {
+            'Authorization': `${auth.token}`
+          }
+        });
+        setProfesores(response.data.usuarios);
+      } catch (error) {
+        console.error(`Hubo un error al obtener los usuarios: ${error}`);
+      }
+    };
+    const eliminarUsuario = async (id) => {
+    if (window.confirm("¿Estás seguro de que quieres eliminar al profesor?")) {
+      try {
+          await axios.delete(`http://localhost:3000/usuarios/${id}`, {
               headers: {
                   'Authorization': `${auth.token}`
               }
-          });
+          }).then((resp)=>{
+            if(resp?.data?.ok){
+              obtenerUsuarios()
+            }
+          })
           // Aquí puedes manejar la respuesta después de eliminar el usuario.
       } catch (error) {
           console.error(`Hubo un error al eliminar el usuario: ${error}`);
       }
-    };
-
+    }
+  };
     const columns = [
         {
             name: 'Nombre y apellido',
@@ -78,26 +101,13 @@ const Profesores = () => {
         </>),
       }
     ];
-    useEffect(() => {
-        const obtenerUsuarios = async () => {
-          try {
-            const response = await axios.get('http://localhost:3000/usuarios?rol=PROFESOR', {
-              headers: {
-                'Authorization': `${auth.token}`
-              }
-            });
-            console.log(response)
-            setProfesores(response.data.usuarios);
-          } catch (error) {
-            console.error(`Hubo un error al obtener los usuarios: ${error}`);
-          }
-        };
 
+    useEffect(() => {
         obtenerUsuarios();
     }, []);
-  console.log(profesores)
   return (
         <PageContainer title={"Profesores"} btnAdd={'/crear-profesor'}>
+          <ToastBootstrap show={showToast} toggleShow={setShowToast} toastMessage={toastMessage} />
           <Table columns={columns} data={profesores} placeholder={"Filtrar por nombre"}/>
         </PageContainer>
   )
